move handle_outlines into shapes beside display_outlines
