+++ src/components/shapes/display_outlines.js
@@ -17,7 +17,7 @@ var i090 = constants.i090;
 var i180 = constants.i180;
 var i270 = constants.i270;
 
-var handleOutline = require('../selections/handle_outline');
+var handleOutline = require('./handle_outline');
 var clearOutlineControllers = handleOutline.clearOutlineControllers;
 
 var helpers = require('./draw_newshape/helpers');
