+++ src/components/shapes/handle_outline.js
